+++ NEWS.md
@@ -1,3 +1,5 @@
+# parsnip (development version)
+
 # parsnip 1.0.3
 
 * Adds documentation and tuning infrastructure for the new `flexsurvspline` engine for the `survival_reg()` model specification from the `censored` package (@mattwarkentin, #831).