+++ NEWS.md
@@ -1,6 +1,6 @@
 # parsnip (development version)
 
-* For censored regression models, a "reverse Kaplan-Meier" curve is computed. This can be used for later use when evaluating this type of model. 
+* For censored regression models, a "reverse Kaplan-Meier" curve is computed. This can be used when evaluating this type of model. 
 
 # parsnip 1.0.3