+++ NEWS.md
@@ -1,6 +1,6 @@
 # parsnip (development version)
 
-* For censored regression models, a "reverse Kaplan-Meier" curve is computed. This can be used when evaluating this type of model. 
+* For censored regression models, a "reverse Kaplan-Meier" curve is computed for the censoring distribution. This can be used when evaluating this type of model (#855).
 
 # parsnip 1.0.3