+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * For censored regression models, a "reverse Kaplan-Meier" curve is computed for the censoring distribution. This can be used when evaluating this type of model (#855).
 
+* The model specification methods for `generics::tune_args()` and 
+  `generics::tunable()` are now registered unconditionally (tidymodels/tune#192).
+
 # parsnip 1.0.3
 
 * Adds documentation and tuning infrastructure for the new `flexsurvspline` engine for the `survival_reg()` model specification from the `censored` package (@mattwarkentin, #831).