+++ NEWS.md
@@ -1,4 +1,4 @@
-# parsnip (development version)
+# parsnip 1.0.4
 
 * For censored regression models, a "reverse Kaplan-Meier" curve is computed for the censoring distribution. This can be used when evaluating this type of model (#855).